+++ packages/press-ui/src/index.js
@@ -45,6 +45,7 @@ import Image from './packages/press-image/press-image.vue';
 import IndexAnchor from './packages/press-index-anchor/press-index-anchor.vue';
 import IndexBar from './packages/press-index-bar/press-index-bar.vue';
 import Info from './packages/press-info/press-info.vue';
+import Keyboard from './packages/press-number-keyboard/press-number-keyboard.vue';
 import Link from './packages/press-link/press-link.vue';
 import List from './packages/press-list/press-list.vue';
 import LoadingPlus from './packages/press-loading-plus/press-loading-plus.vue';
@@ -149,6 +150,7 @@ const components = {
   IndexAnchor,
   IndexBar,
   Info,
+  Keyboard,
   Link,
   List,
   LoadingPlus,