+++ packages/press-ui/src/packages/press-picker-plus/tests/demo.spec.ts
@@ -1,0 +1,5 @@
+import Demo from '../demo.vue';
+import { snapshotDemo } from '../../../../tests/unit/demo';
+
+
+snapshotDemo(Demo);